+++ unittest/common/mem_pool_test.cpp
@@ -57,6 +57,53 @@ TEST(test_mem_pool_item, test_mem_pool_item_basic)
   ASSERT_EQ(pool_size, mem_pool_item.get_size());
 }
 
+struct Frame
+{
+  char buf[4096];
+
+  void reinit() {}
+  void reset() {}
+};
+
+// PROCESS WILL CRASH!
+TEST(mm, DISABLED_mm_illegal_access)
+{
+  using namespace common;
+  MemPoolSimple<Frame> pool{"mm_illegal_access"};
+  ASSERT_TRUE(pool.init() == 0);
+  auto frame = pool.alloc();
+
+  // Free frame
+  pool.free(frame);
+
+  // Access frame. Process WILL CRASH!
+  auto buf = frame->buf;
+  buf[0]   = '1';
+}
+
+TEST(mm, mm_legal_access)
+{
+  using namespace common;
+  MemPoolSimple<Frame> pool{"mm_illegal_access"};
+  ASSERT_TRUE(pool.init(false, 3, 3) == 0);
+  std::vector<Frame *> frames;
+
+  for (auto i = 0; i < 3; ++i) {
+    frames.push_back(pool.alloc());
+  }
+
+  for (auto i = 0; i < 3; ++i) {
+    pool.free(frames.at(i));
+  }
+
+  frames.clear();
+  for (auto i = 0; i < 3; ++i) {
+    Frame* f = pool.alloc();
+    f->buf[0] = 1;
+    pool.free(f);
+  }
+}
+
 int main(int argc, char **argv)
 {
